Restore master.AcknowledgedResponse
