+++ server/src/main/java/org/opensearch/action/admin/cluster/settings/ClusterUpdateSettingsResponse.java
@@ -32,7 +32,6 @@
 
 package org.opensearch.action.admin.cluster.settings;
 
-import org.opensearch.action.support.master.AcknowledgedResponse;
 import org.opensearch.common.ParseField;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
