+++ server/src/main/java/org/opensearch/action/support/clustermanager/AcknowledgedRequestBuilder.java
@@ -32,6 +32,7 @@
 package org.opensearch.action.support.clustermanager;
 
 import org.opensearch.action.ActionType;
+import org.opensearch.action.support.master.AcknowledgedResponse;
 import org.opensearch.client.OpenSearchClient;
 import org.opensearch.common.unit.TimeValue;
